requirements.txt
# IoT Face Detection & Photo Capture System
# Minimal dependencies - no facial recognition library needed

# Core Vision
opencv-python==4.8.1.78
numpy==1.24.3

# QR Code Support
pyzbar==0.1.9
qrcode==7.4.2
pillow==10.0.0

# Database Support (SQLite - no server needed)
# SQLite is built into Python, no additional package needed

# Configuration & Utilities
python-dotenv==1.0.0
pydantic==2.4.2

# Cloud Sync
supabase==2.3.0
requests==2.31.0

# GPIO/Hardware Control
RPi.GPIO==0.7.1

# Networking
netifaces==0.11.0

# Optional: Testing
pytest==7.4.3
pytest-cov==4.1.0
